move messageStore2 to messageStore
