+++ src/modules/MessageStore/messageStoreHelper.js
@@ -1,168 +1,266 @@
 import * as messageHelper from '../../lib/messageHelper';
+import removeUri from '../../lib/removeUri';
+import syncTypes from '../../enums/syncTypes';
 
-function pushOrReplaceOrDeleteMessage({
-  messages,
-  message,
-  isFind,
-  replaceMessage,
-  pushMessage,
-  deleteMessage
-}) {
-  const messageLength = messages.length;
-  let messageExistIndex = null;
-  if (messageLength > 0) {
-    for (let index = (messageLength - 1); index >= 0; index -= 1) {
-      if (isFind({
-        oldMessage: messages[index],
-        newMessage: message,
-      })) {
-        messageExistIndex = index;
-        break;
-      }
-    }
-  }
-  if (messageExistIndex === null) {
-    if (messageHelper.messageIsAcceptable(message)) {
-      pushMessage(message);
-    }
-    return;
-  }
-  if (messageHelper.messageIsDeleted(message)) {
-    if (messages[messageExistIndex].id === message.id) {
-      deleteMessage(messageExistIndex);
-    }
-    return;
-  }
-  replaceMessage({ index: messageExistIndex, newMessage: message });
-}
-
-export function pushMessageToConversationMessages({ messages, message }) {
-  const isFind = ({ oldMessage, newMessage }) => (
-    oldMessage.id === newMessage.id
-  );
-  const replaceMessage = ({ index, newMessage }) => {
-    messages[index] = newMessage;
-  };
-  const pushMessage = (newMessage) => {
-    messages.push(newMessage);
-  };
-  const deleteMessage = (index) => {
-    messages.splice(index, 1);
-  };
-  pushOrReplaceOrDeleteMessage({
-    messages,
-    message,
-    isFind,
-    replaceMessage,
-    pushMessage,
-    deleteMessage
-  });
-  return messages;
-}
-
-export function pushMessageToMesages({ messages, message }) {
-  const isFind = ({ oldMessage, newMessage }) => (
-    oldMessage.id === newMessage.id ||
-      oldMessage.conversation.id === newMessage.conversation.id
-  );
-  const replaceMessage = ({ index, newMessage }) => {
-    const oldCreated = new Date(messages[index].creationTime);
-    const newCreated = new Date(message.creationTime);
-    if (newCreated >= oldCreated) {
-      messages.splice(index, 1);
-      messages.push(newMessage);
-    }
-  };
-  const pushMessage = (newMessage) => {
-    messages.push(newMessage);
-  };
-  const deleteMessage = (index) => {
-    messages.splice(index, 1);
-  };
-
-  pushOrReplaceOrDeleteMessage({
-    messages,
-    message,
-    isFind,
-    replaceMessage,
-    pushMessage,
-    deleteMessage
-  });
-  return messages;
-}
-
-export function getNewConversationsAndMessagesFromRecords({
-  records,
-  syncToken,
-  conversations,
-  messages,
-}) {
-  records.forEach((record) => {
-    if (!messageHelper.messaageIsTextMessage(record)) {
-      return;
-    }
-    const conversationId = record.conversation.id;
-    let conversation = conversations[conversationId];
-    if (!conversation) {
-      conversation = { messages: [] };
-    }
-    const oldMessages = conversation.messages;
-    conversation.messages = pushMessageToConversationMessages({
-      messages: oldMessages,
-      message: record,
-    });
-    if (syncToken) {
-      conversation.syncToken = syncToken;
-    }
-    conversation.id = conversationId;
-    conversations[conversationId] = conversation;
-    messages = pushMessageToMesages({
-      messages,
-      message: record
-    });
-  });
-  return { conversations, messages };
-}
-
-export function filterConversationUnreadMessages(conversation) {
-  const unReadMessages = conversation.messages.filter(record => (
-    record.direction === 'Inbound' &&
-    !(messageHelper.messageIsDeleted(record)) &&
-    record.readStatus !== 'Read'
-  ));
-  return unReadMessages;
-}
-
-export function updateMessagesUnreadCounts(messages, conversations) {
-  let totalUnreadCounts = 0;
-  for (let index = 0; index < messages.length; index += 1) {
-    const message = messages[index];
-    const conversation = conversations[message.conversation.id];
-    const unReadMessages = filterConversationUnreadMessages(conversation);
-    totalUnreadCounts += unReadMessages.length;
-    message.isRead = (unReadMessages.length === 0);
-  }
+export function normalizeRecord(record) {
   return {
-    messages,
-    unreadCounts: totalUnreadCounts
+    ...record,
+    conversationId: record.conversation.id,
   };
+}
+
+export function messageIsUnread(message) {
+  return (
+    message.direction === 'Inbound' &&
+    message.readStatus !== 'Read' &&
+    !(messageHelper.messageIsDeleted(message))
+  );
 }
 
 export function getMessageSyncParams({ syncToken, conversationId }) {
   if (syncToken) {
     return {
       syncToken,
-      syncType: 'ISync',
+      syncType: syncTypes.iSync,
     };
   }
   const lastSevenDate = new Date();
   lastSevenDate.setDate(lastSevenDate.getDate() - 7);
   const params = {
-    syncType: 'FSync',
+    syncType: syncTypes.fSync,
     dateFrom: lastSevenDate.toISOString(),
   };
   if (conversationId) {
     params.conversationId = conversationId;
   }
   return params;
+}
+
+export function prepareNewMessagesData({
+  messages,
+  conversations,
+  conversationMap,
+  syncToken,
+  syncConversationId,
+}) {
+  const newConversations = [];
+  const newConversationMap = {};
+  const newMessages = [];
+  const messageMap = {};
+  // copy old conversationMap to new conversationMap hash
+  Object.keys(conversationMap).forEach((key) => {
+    const conversation = {
+      ...conversationMap[key],
+      unreadMessages: {
+        ...conversationMap[key].unreadMessages,
+      },
+    };
+    // if converstation is not sync with conversation Id, update all conversation sync token
+    if (syncToken && !syncConversationId) {
+      conversation.syncToken = syncToken;
+    }
+    newConversationMap[key] = conversation;
+  });
+  // copy old conversations to new conversations array
+  conversations.forEach((conversation) => {
+    newConversations.push({ ...conversation });
+  });
+  messages.forEach((message) => {
+    newMessages.push({ ...message });
+    messageMap[message.id] = newMessages.length - 1;
+  });
+  return {
+    newConversations,
+    newConversationMap,
+    newMessages,
+    messageMap,
+  };
+}
+
+export function findIndexOfConversations(newConversationMap, record) {
+  const conversationId = record.conversation && record.conversation.id;
+  const existedIndex =
+    newConversationMap[conversationId] &&
+    newConversationMap[conversationId].index;
+  if (existedIndex !== undefined && existedIndex !== null) {
+    return existedIndex;
+  }
+  return -1;
+}
+
+export function findIndexOfMessages(messageMap, record) {
+  if (messageMap[record.id] !== undefined) {
+    return messageMap[record.id];
+  }
+  return -1;
+}
+
+export function calcUnreadCount(conversation) {
+  return Object.keys(conversation.unreadMessages).length;
+}
+
+export function pushRecordsToMessageData({
+  messages,
+  conversations,
+  conversationMap,
+  records,
+  syncToken,
+  syncConversationId,
+}) {
+  const {
+    newConversations,
+    newConversationMap,
+    newMessages,
+    messageMap,
+  } = prepareNewMessagesData({
+    messages,
+    conversations,
+    conversationMap,
+    syncToken,
+    syncConversationId,
+  });
+  const addMessageToMessageMap = (message, index) => {
+    messageMap[message.id] = index;
+  };
+  const setSyncTokenToConversation = (conversation) => {
+    if (
+      syncToken &&
+      (
+        !syncConversationId ||
+        (syncConversationId && syncConversationId === conversation.id)
+      )
+    ) {
+      conversation.syncToken = syncToken;
+    }
+  };
+  const addMessageToConversationMap = (message, index) => {
+    const conversationId = message.conversationId;
+    if (conversationId) {
+      const conversation = newConversationMap[conversationId] || { unreadMessages: {} };
+      conversation.index = index;
+      conversation.id = conversationId;
+      setSyncTokenToConversation(conversation);
+      if (messageIsUnread(message)) {
+        conversation.unreadMessages[message.id] = 1;
+      } else if (conversation.unreadMessages[message.id]) {
+        delete conversation.unreadMessages[message.id];
+      }
+      newConversationMap[conversationId] = conversation;
+    }
+  };
+  const pushMessageToConversations = (record) => {
+    const message = normalizeRecord(removeUri(record));
+    const index = newConversations.length;
+    addMessageToConversationMap(message, index);
+    const conversation = newConversationMap[message.conversationId];
+    if (conversation) {
+      message.unreadCounts = calcUnreadCount(conversation);
+    } else {
+      message.unreadCounts = 0;
+    }
+    newConversations.push(message);
+  };
+  const pushMessageToMessages = (record) => {
+    const message = normalizeRecord(removeUri(record));
+    newMessages.push(message);
+    addMessageToMessageMap(message, newMessages.length - 1);
+  };
+  // TODO: delete message or conversation?
+  const deleteMessageFromConversations = (index, record) => {
+    const message = newConversations[index];
+    if (message.id === record.id) {
+      const conversationMessages = newMessages.filter(oldMessage => (
+        oldMessage.id !== message.id && oldMessage.conversationId === message.conversationId
+      ));
+      if (conversationMessages.length === 0) {
+        newConversations.splice(index, 1);
+        delete newConversationMap[record.conversation.id];
+        return;
+      }
+      newConversations[index] = conversationMessages[conversationMessages.length - 1];
+    }
+    const conversation = newConversationMap[record.conversation.id];
+    setSyncTokenToConversation(conversation);
+    delete conversation.unreadMessages[record.id];
+    message.unreadCounts = calcUnreadCount(conversation);
+  };
+  const deleteMessageFromMessages = (index, record) => {
+    newMessages.splice(index, 1);
+    delete messageMap[record.id];
+  };
+  const replaceMessageInConversations = (index, record) => {
+    const oldConversation = newConversations[index];
+    const newMessage = {
+      ...oldConversation,
+      ...normalizeRecord(removeUri(record)),
+    };
+    const oldCreated = new Date(oldConversation.creationTime);
+    const newCreated = new Date(record.creationTime);
+    if (newCreated >= oldCreated) {
+      // move the message to the top of new Messages
+      newConversations.splice(index, 1);
+      newConversations.push(newMessage);
+      addMessageToConversationMap(newMessage, newConversations.length - 1);
+    } else {
+      addMessageToConversationMap(newMessage, index);
+    }
+    const conversation = newConversationMap[newMessage.conversationId];
+    newMessage.unreadCounts = calcUnreadCount(conversation);
+  };
+  const replaceMessageInMessages = (index, record) => {
+    newMessages[index] = normalizeRecord(removeUri(record));
+  };
+  records.forEach((record) => {
+    if (!record.conversation) {
+      return;
+    }
+    const existedIndexofMessages = findIndexOfMessages(messageMap, record);
+    const existedIndexofConversations = findIndexOfConversations(newConversationMap, record);
+    const isDeleted = messageHelper.messageIsDeleted(record);
+    const isAcceptable = messageHelper.messageIsAcceptable(record);
+    if (existedIndexofMessages > -1) {
+      if (isDeleted) {
+        deleteMessageFromMessages(existedIndexofMessages, record);
+      } else {
+        replaceMessageInMessages(existedIndexofMessages, record);
+      }
+    } else if (isAcceptable) {
+      pushMessageToMessages(record);
+    }
+    if (existedIndexofConversations > -1) {
+      if (isDeleted) {
+        deleteMessageFromConversations(existedIndexofConversations, record);
+      } else {
+        replaceMessageInConversations(existedIndexofConversations, record);
+      }
+    } else if (isAcceptable) {
+      pushMessageToConversations(record);
+    }
+  });
+  return {
+    conversations: newConversations,
+    conversationMap: newConversationMap,
+    messages: newMessages,
+  };
+}
+
+export function updateConversationRecipients({
+  messages,
+  conversations,
+  conversationMap,
+  conversationId,
+  recipients,
+}) {
+  const newConversations = [];
+  conversations.forEach((conversation) => {
+    newConversations.push({ ...conversation });
+  });
+  const conversationIndex = conversationMap[conversationId].index;
+  const conversation = newConversations[conversationIndex];
+  conversation.recipients = recipients;
+  return {
+    messages,
+    conversationMap,
+    conversations: newConversations,
+  };
 }
