+++ src/modules/MessageStore/messageStoreStatus.js
@@ -2,6 +2,5 @@ import Enum from '../../lib/Enum';
 
 export default new Enum([
   'syncing',
-  'pushing',
   'idle',
 ], 'messageStore');
