+++ src/modules/MessageStoreOld/messageStoreStatus.js
@@ -2,5 +2,6 @@ import Enum from '../../lib/Enum';
 
 export default new Enum([
   'syncing',
+  'pushing',
   'idle',
 ], 'messageStore');
